requirements.txt
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
opencv-python==4.9.0.80
tensorflow
scikit-learn==1.4.2
Keras-Preprocessing==1.1.2